requirements/base.txt
Django==5.1
django-filter==24.3
django-split-settings==1.3.2
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
python-decouple==3.8
